Working with test code still in, need to fix lengths and setup key changing in stub

+++ unpacker_funcs/unpacker_funcs.h
@@ -1,2 +1,2 @@
-int getDataPtr(char *, long, int);
+int getDataPtr(char *, long, long);
